+++ include/bouncer.hpp
@@ -1,16 +1,10 @@
 #pragma once
 
-#include <string>
-
-class Bouncer {
-public:
-  Bouncer();
-
-  static std::string executeCommand(const char *cmd);
-  static bool isProcessRunning(const std::string &processName);
-
-  static bool isHyprlandRunning();
-  static bool isHyprpaperRunning();
-  static bool isHyprpaperIpcEnabled();
-  static bool doesZenityRun();
-};
+namespace Bouncer {
+  /*
+   * Checks if all requirements are met to run the program
+     * If not, exits the program with an error message
+     * Error also logged in log file if in production
+   */
+  void checkRequirements();
+} // namespace Bouncer